Refactor the folder structure
