--- src/components/hero/slider.css
@@ -1,41 +1,0 @@
-.slider {
-  margin-top: 50px;
-}
-.slider img {
-  width: 100%;
-}
-
-/*  order */
-.order {
-  margin-top: 50px;
-}
-.order .container {
-  padding: 50px 30px;
-  background: #fff;
-}
-.order .num h1 {
-  width: 70px;
-  height: 70px;
-  line-height: 70px;
-  border-radius: 50%;
-  background: #feeec8;
-  text-align: center;
-  margin-right: 20px;
-}
-.order h3 {
-  font-weight: 600;
-  margin-bottom: 10px;
-}
-.order p {
-  font-size: 17px;
-  line-height: 25px;
-}
-.order .box:nth-child(2) h1 {
-  background: #ceeffe;
-}
-.order .box:nth-child(3) h1 {
-  background: #d4f8c4;
-}
-.order .box:nth-child(4) h1 {
-  background: #d8dafe;
-}
