+++ src/pages/home/components/category/Category.jsx
@@ -5,69 +5,7 @@ import Slider from "react-slick"
 import "slick-carousel/slick/slick-theme.css"
 import "slick-carousel/slick/slick.css"
 import "./category.css"
-
-const category = [
-  {
-    id: 1,
-    title: "face & skin care",
-    cover: "../images/category/ca1.png",
-  },
-  {
-    id: 2,
-    title: "baby care",
-    cover: "../images/category/ca2.png",
-  },
-  {
-    id: 3,
-    title: "beauty care",
-    cover: "../images/category/ca3.png",
-  },
-  {
-    id: 4,
-    title: "contraceptive",
-    cover: "../images/category/ca4.png",
-  },
-  {
-    id: 5,
-    title: "feminine hygiene",
-    cover: "../images/category/ca5.png",
-  },
-  {
-    id: 6,
-    title: "health & protein",
-    cover: "../images/category/ca6.png",
-  },
-  {
-    id: 7,
-    title: "health & wellness",
-    cover: "../images/category/ca7.png",
-  },
-  {
-    id: 8,
-    title: "medicine & first aid",
-    cover: "../images/category/ca8.png",
-  },
-  {
-    id: 9,
-    title: "oral care",
-    cover: "../images/category/ca9.png",
-  },
-  {
-    id: 10,
-    title: "pregnacy",
-    cover: "../images/category/ca10.png",
-  },
-  {
-    id: 11,
-    title: "sexual well-being",
-    cover: "../images/category/ca11.png",
-  },
-  {
-    id: 12,
-    title: "Household",
-    cover: "../images/category/ca12.jpg",
-  },
-]
+import { category } from "../../../../data/categoryData"
 
 function SampleNextArrow(props) {
   const { onClick } = props
