+++ src/pages/home/components/category/category.css
